backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.12
uvicorn[standard]==0.34.3

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication and Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Environment and Configuration
python-dotenv==1.1.0
pydantic-settings==2.9.1

# HTTP client
httpx==0.28.1

# Validation and serialization
pydantic[email]==2.11.5
email-validator==2.2.0

# Logging
structlog==25.4.0

# Testing
pytest==8.4.0
pytest-asyncio==1.0.0

# Development
black==25.1.0
flake8==7.2.0
isort==6.0.1
pre-commit==4.2.0
